Make sure that byte array refs for variables are deleted when deleting a case instance

+++ modules/flowable-cmmn-engine/src/main/java/org/flowable/cmmn/engine/impl/persistence/entity/CaseInstanceEntityManagerImpl.java
@@ -31,8 +31,8 @@ import org.flowable.common.engine.impl.persistence.entity.AbstractEngineEntityMa
 import org.flowable.entitylink.service.impl.persistence.entity.EntityLinkEntityManager;
 import org.flowable.eventsubscription.service.EventSubscriptionService;
 import org.flowable.identitylink.service.impl.persistence.entity.IdentityLinkEntityManager;
-import org.flowable.job.api.Job;
 import org.flowable.job.api.ExternalWorkerJob;
+import org.flowable.job.api.Job;
 import org.flowable.job.service.impl.DeadLetterJobQueryImpl;
 import org.flowable.job.service.impl.ExternalWorkerJobQueryImpl;
 import org.flowable.job.service.impl.JobQueryImpl;
@@ -45,6 +45,7 @@ import org.flowable.job.service.impl.persistence.entity.SuspendedJobEntityManage
 import org.flowable.job.service.impl.persistence.entity.TimerJobEntityManager;
 import org.flowable.task.service.impl.persistence.entity.TaskEntity;
 import org.flowable.task.service.impl.persistence.entity.TaskEntityManager;
+import org.flowable.variable.service.impl.persistence.entity.VariableInstanceEntity;
 import org.flowable.variable.service.impl.persistence.entity.VariableInstanceEntityManager;
 
 /**
@@ -88,8 +89,24 @@ public class CaseInstanceEntityManagerImpl
         CaseInstanceEntity caseInstanceEntity = dataManager.findById(caseInstanceId);
 
         // Variables
-        getVariableInstanceEntityManager().deleteByScopeIdAndScopeTypes(caseInstanceId, ScopeTypes.CMMN_DEPENDENT);
-        
+        // variables can have byte array refs, so fetch them and delete the byte array refs if needed
+        List<VariableInstanceEntity> variableInstances = getVariableInstanceEntityManager()
+                .createInternalVariableInstanceQuery()
+                .scopeId(caseInstanceEntity.getId())
+                .scopeTypes(ScopeTypes.CMMN_DEPENDENT)
+                .list();
+        boolean deleteVariableInstances = !variableInstances.isEmpty();
+
+        for (VariableInstanceEntity variableInstance : variableInstances) {
+            if (variableInstance.getByteArrayRef() != null && variableInstance.getByteArrayRef().getId() != null) {
+                variableInstance.getByteArrayRef().delete();
+            }
+        }
+
+        if (deleteVariableInstances) {
+            getVariableInstanceEntityManager().deleteByScopeIdAndScopeTypes(caseInstanceId, ScopeTypes.CMMN_DEPENDENT);
+        }
+
         // Identity links
         getIdentityLinkEntityManager().deleteIdentityLinksByScopeIdAndScopeType(caseInstanceId, ScopeTypes.CMMN);
